+++ src/test/java/guitests/guihandles/event/EventListPanelHandle.java
@@ -1,8 +1,10 @@
-package guitests.guihandles;
+package guitests.guihandles.event;
 
 import java.util.List;
 import java.util.Optional;
 
+import guitests.guihandles.NodeHandle;
+import guitests.guihandles.event.EventCardHandle;
 import javafx.scene.control.ListView;
 import seedu.address.model.event.ReadOnlyEvent;
 import seedu.address.ui.event.EventCard;
