+++ src/test/java/guitests/guihandles/person/PersonCardHandle.java
@@ -1,8 +1,9 @@
-package guitests.guihandles;
+package guitests.guihandles.person;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
+import guitests.guihandles.NodeHandle;
 import javafx.scene.Node;
 import javafx.scene.control.Label;
 import javafx.scene.layout.Region;
