+++ src/test/java/guitests/guihandles/person/PersonListPanelHandle.java
@@ -1,8 +1,10 @@
-package guitests.guihandles;
+package guitests.guihandles.person;
 
 import java.util.List;
 import java.util.Optional;
 
+import guitests.guihandles.NodeHandle;
+import guitests.guihandles.person.PersonCardHandle;
 import javafx.scene.control.ListView;
 import seedu.address.model.person.ReadOnlyPerson;
 import seedu.address.ui.person.PersonCard;
